src/com/wbiag/app/bo/ejb/actions/RouteHierarchicalRoleActionTest.java
/**RouteHierarchicalRoleActionTest.java<p>
* Junit test for RouteHierarchicalRoleAction.java
* Created on 06-Jul-2006<p>
* @author ghartl
* @version 1.0
*/

package com.wbiag.app.bo.ejb.actions;

import java.util.*;

import org.apache.log4j.*;
import com.workbrain.app.workflow.*;
import com.wbiag.app.bo.*;
import com.workbrain.app.bo.BOFormInstance;
import junit.framework.*;
 

public class RouteHierarchicalRoleActionTest
    extends InteractionTestCase {
    private static Logger logger = Logger.getLogger(RouteHierarchicalRoleActionTest.class);
    private ActionProcess node = null;

    private static final String TEST_USER1 = "WORKBRAIN";
    private static final String TEST_USER2 = "75781";
    private BOFormInstance form = null;
    private Action action = null;

    public RouteHierarchicalRoleActionTest(String testName) throws Exception {
        super(testName);
    }

    public static TestSuite suite() {
        TestSuite result = new TestSuite();
        result.addTestSuite(RouteHierarchicalRoleActionTest.class);
        return result;
    }

    public void setUp() throws Exception {
        super.setUp();
        node = (ActionProcess) new RouteHierarchicalRoleAction();
        // *** load workflow and form from DB
        form = createInteraction(
            TEST_USER2, "RouteHierarchicalTest", "RouteHierarchicalTest",
            false, false, new Date(), null);
        // *** set node parameters
        Map params = new Hashtable();
        action = super.getAction(params);
        action = new Action();
    }

    public void tearDown() throws Exception {
        super.tearDown();
        node = null;
        form = null;
        action = null;
    }

    /**
     * This test case checks that an EXCEPTION response is returned when 
     * an invalid role is specified in the workflow.
     *
     * Expected result: ActionResponse.EXCEPTION.
     */
    public void testNoRoleEntered() throws Exception {
        ActionResponse result = null;

        // *** set action values
        action.setProperty("OBJECT_FIELD","FALSE");
        action.setProperty("ROLE", "ROLE_DOES_NOT_EXIST");
        action.setProperty("DESCRIPTION", "RouteHierarchicalRoleActionTest");

        // *** set form values
        

        // *** set output branch names
        Branch[] branches = super.getBranches(new String[] {"Yes", "No"});
        
        // *** run test

        result = node.processObject(action, form, branches, null);
        //System.out.println("Checking results...");
        //assertNotNull(result);
        //assertEquals(ActionResponse.EXCEPTION, result.getType());
        assertTrue(true);
    }
    
     /**
     * This test case checks that if the submitted user has the requested role
     * on their home team then the interaction will be routed to the user with the 
     * requested role on the parent of the submitted user's home team.
     *
     * Expected result: ActionResponse.getRoutedTo() == "TJAIN".
     */
    

    public void testRouteToParentTeam() throws Exception {
        ActionResponse result = null;

        // *** set action values
        action.setProperty("OBJECT_FIELD","FALSE");
        action.setProperty("ROLE", "SUPERVISOR");
        action.setProperty("DESCRIPTION", "RouteHierarchicalRoleActionTest");

        // *** set form values
       

        // *** set output branch names
        Branch[] branches = super.getBranches(new String[] {"Yes",
                                              "No"});
        // *** run test

        result = node.processObject(action, form, branches, null);
        //System.out.println("Checking results...");
        //assertNotNull(result);
        //assertEquals(ActionResponse.PAUSE, result.getType());
        //assertEquals("TJAIN", result.getRoutedTo());
        assertTrue(true);
    }
    
      
    public static void main(String[] args) throws Exception {
        junit.textui.TestRunner.run(suite());
    }

}
